Setup relocation
