+++ test/fixtures/active_record.rb
@@ -908,6 +908,11 @@ module Api
     class NumerosTelefoneController < JSONAPI::ResourceController
     end
   end
+
+  module CustomActions
+    class PostsController < JSONAPI::ResourceController
+    end
+  end
 end
 
 module Api
@@ -1738,6 +1743,77 @@ module Api
   module V8
     class NumeroTelefoneResource < JSONAPI::Resource
       attribute :numero_telefone
+    end
+  end
+
+  module CustomActions
+    class CommentResource < Api::V1::CommentResource; end;
+    class PersonResource < Api::V1::PersonResource; end;
+    class PostResource < Api::V1::PostResource
+      custom_action :favorite, type: :get, apply: (lambda do |model, context, data| # you can use apply with lambda
+        Post.first
+      end)
+
+      custom_action :draft, type: :post # without apply, it will call a method with correponsing name
+
+      def draft(data)
+        spawn_post
+      end
+
+      custom_action :draft_with_comments, type: :post, includes: 'comments'
+
+      def draft_with_comments(data)
+        spawn_post
+      end
+
+      custom_action :nil_response
+
+      def nil_response(data)
+        nil
+      end
+
+      custom_action :invalid_model
+
+      def invalid_model(data)
+        Post.create
+      end
+
+      custom_action :last_author
+
+      def last_author(data)
+        Person.last
+      end
+
+      custom_action :last, level: :collection
+
+      def last(data)
+        Post.last
+      end
+
+      custom_action :friend, method: 'my_friend'
+
+      def my_friend(data)
+        Post.last
+      end
+
+      custom_action :custom_draft, type: :put
+
+      def custom_draft(data)
+        title = data.require(:attributes).permit(:title)[:title] # strong params keeps you safer
+        Post.create(
+          title: title,
+          author: Person.create(name: 'Test', date_joined: 1.days.ago)
+        )
+      end
+
+      private
+
+      def spawn_post
+        Post.create(
+          title: 'Custom action post',
+          author: Person.create(name: 'Test', date_joined: 1.days.ago)
+        )
+      end
     end
   end
 end
